moved selectors utils
