+++ src/app/content/utils/selectorsUtils.ts
