+++ src/app/content/utils/sharedHighlightsUtils/index.ts
@@ -6,7 +6,7 @@ import { CountsPerSource } from '../../types';
 import createSummaryHighlightsLoader from './createSummaryHighlightsLoader';
 import extractDataFromHighlightClientResponse from './extractDataFromHighlightClientResponse';
 import getHighlightLocationFilterForPage from './getHighlightLocationFilterForPage';
-import loadAllHighlights from './loadAllHighlights'
+import loadAllHighlights from './loadAllHighlights';
 
 const extractTotalCounts = (countsPerSource: CountsPerSource) =>
   mapValues(pickBy<CountsPerSource>(isDefined), countsPerSource);
